🚚 Move utilities to CmHelpers, don't auto-complete mentions in code blocks or html blocks
